+++ src/bus/dram.rs
@@ -34,7 +34,7 @@ impl Dram {
         let dram_size = new_dram.len();
         Dram {
             dram: new_dram,
-            base_addr: virt_entry as u32,
+            base_addr: virt_entry,
             size: dram_size,
         }
     }
@@ -44,7 +44,7 @@ impl Dram {
 impl Device for Dram {
     // is addr in device address space
     fn in_range(&self, addr: u64) -> bool {
-        (self.base_addr..=self.base_addr + self.size as u32).contains(&addr)
+        (self.base_addr..=self.base_addr + self.size as u64).contains(&addr)
     }
 
     // address to raw index
@@ -59,20 +59,20 @@ impl Device for Dram {
     }
 
     // store
-    fn store8(&mut self, addr: u64, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store8(&mut self, addr: u64, data: u64) -> Result<(), (Option<u64>, TrapCause, String)> {
         let index = self.addr2index(addr);
         self.dram[index] = (data & 0xFF) as u8;
         Ok(())
     }
 
-    fn store16(&mut self, addr: u64, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store16(&mut self, addr: u64, data: u64) -> Result<(), (Option<u64>, TrapCause, String)> {
         let index = self.addr2index(addr);
         self.dram[index + 1] = ((data >> 8) & 0xFF) as u8;
         self.dram[index + 0] = ((data >> 0) & 0xFF) as u8;
         Ok(())
     }
 
-    fn store32(&mut self, addr: u64, data: u32) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store32(&mut self, addr: u64, data: u64) -> Result<(), (Option<u64>, TrapCause, String)> {
         let index = self.addr2index(addr);
         self.dram[index + 3] = ((data >> 24) & 0xFF) as u8;
         self.dram[index + 2] = ((data >> 16) & 0xFF) as u8;
@@ -81,7 +81,7 @@ impl Device for Dram {
         Ok(())
     }
 
-    fn store64(&mut self, addr: u64, data: i64) -> Result<(), (Option<u32>, TrapCause, String)> {
+    fn store64(&mut self, addr: u64, data: u64) -> Result<(), (Option<u64>, TrapCause, String)> {
         let index = self.addr2index(addr);
         self.dram[index + 7] = ((data >> 56) & 0xFF) as u8;
         self.dram[index + 6] = ((data >> 48) & 0xFF) as u8;
@@ -95,25 +95,25 @@ impl Device for Dram {
     }
 
     // load
-    fn load8(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load8(&self, addr: u64) -> Result<u64, (Option<u64>, TrapCause, String)> {
         let index = self.addr2index(addr);
-        Ok(self.dram[index] as i8 as i32 as u32)
+        Ok(self.dram[index] as i8 as i32 as u64)
     }
 
-    fn load16(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load16(&self, addr: u64) -> Result<u64, (Option<u64>, TrapCause, String)> {
         let index = self.addr2index(addr);
-        Ok(((self.dram[index + 1] as i16) << 8 | (self.dram[index + 0] as i16)) as i32 as u32)
+        Ok(((self.dram[index + 1] as i16) << 8 | (self.dram[index + 0] as i16)) as i32 as u64)
     }
 
-    fn load32(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load32(&self, addr: u64) -> Result<u64, (Option<u64>, TrapCause, String)> {
         let index = self.addr2index(addr);
-        Ok((self.dram[index + 3] as u32) << 24
-            | (self.dram[index + 2] as u32) << 16
-            | (self.dram[index + 1] as u32) << 8
-            | (self.dram[index + 0] as u32))
+        Ok((self.dram[index + 3] as u64) << 24
+            | (self.dram[index + 2] as u64) << 16
+            | (self.dram[index + 1] as u64) << 8
+            | (self.dram[index + 0] as u64))
     }
 
-    fn load64(&self, addr: u64) -> Result<u64, (Option<u32>, TrapCause, String)> {
+    fn load64(&self, addr: u64) -> Result<u64, (Option<u64>, TrapCause, String)> {
         let index = self.addr2index(addr);
         Ok((self.dram[index + 7] as u64) << 56
             | (self.dram[index + 6] as u64) << 48
@@ -125,14 +125,14 @@ impl Device for Dram {
             | (self.dram[index + 0] as u64))
     }
 
-    fn load_u8(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load_u8(&self, addr: u64) -> Result<u64, (Option<u64>, TrapCause, String)> {
         let index = self.addr2index(addr);
-        Ok(self.dram[index] as u32)
+        Ok(self.dram[index] as u64)
     }
 
-    fn load_u16(&self, addr: u64) -> Result<u32, (Option<u32>, TrapCause, String)> {
+    fn load_u16(&self, addr: u64) -> Result<u64, (Option<u64>, TrapCause, String)> {
         let index = self.addr2index(addr);
-        Ok(((self.dram[index + 1] as u16) << 8 | (self.dram[index + 0] as u16)) as u32)
+        Ok(((self.dram[index + 1] as u16) << 8 | (self.dram[index + 0] as u16)) as u64)
     }
 }
 
@@ -150,8 +150,8 @@ mod tests {
         };
         let mut addr = 0;
         let mut test_8 = |data: i32| {
-            Dram::store8(dram, addr, data as u32).unwrap();
-            assert_eq!(data as u32, Dram::load_u8(dram, addr).unwrap());
+            Dram::store8(dram, addr, data as u64).unwrap();
+            assert_eq!(data as u64, Dram::load_u8(dram, addr).unwrap());
             addr += 2;
         };
 
@@ -160,9 +160,9 @@ mod tests {
         test_8(0b01111111);
 
         let minus_num = -42;
-        Dram::store8(dram, addr, minus_num as u32).unwrap();
+        Dram::store8(dram, addr, minus_num as u64).unwrap();
         assert_ne!(minus_num, Dram::load_u8(dram, addr).unwrap() as i32);
-        Dram::store16(dram, addr, minus_num as u32).unwrap();
+        Dram::store16(dram, addr, minus_num as u64).unwrap();
         assert_eq!(214, Dram::load_u8(dram, addr).unwrap());
     }
 
@@ -175,8 +175,8 @@ mod tests {
         };
         let mut addr = 0;
         let mut test_8 = |data: i32| {
-            Dram::store8(dram, addr, data as u32).unwrap();
-            assert_eq!(data as u32, Dram::load8(dram, addr).unwrap());
+            Dram::store8(dram, addr, data as u64).unwrap();
+            assert_eq!(data as u64, Dram::load8(dram, addr).unwrap());
             addr += 2;
         };
 
@@ -198,8 +198,8 @@ mod tests {
         };
         let mut addr = 0;
         let mut test_16 = |data: i32| {
-            Dram::store16(dram, addr, data as u32).unwrap();
-            assert_eq!(data as u32, Dram::load16(dram, addr).unwrap());
+            Dram::store16(dram, addr, data as u64).unwrap();
+            assert_eq!(data as u64, Dram::load16(dram, addr).unwrap());
             addr += 2;
         };
 
@@ -222,8 +222,8 @@ mod tests {
         };
         let mut addr = 0;
         let mut test_u16 = |data: i32| {
-            Dram::store16(dram, addr, data as u32).unwrap();
-            assert_eq!(data as u32, Dram::load_u16(dram, addr).unwrap());
+            Dram::store16(dram, addr, data as u64).unwrap();
+            assert_eq!(data as u64, Dram::load_u16(dram, addr).unwrap());
             addr += 2;
         };
 
@@ -233,9 +233,9 @@ mod tests {
         test_u16(0b0111111111111111);
 
         let minus_num = -42;
-        Dram::store16(dram, addr, minus_num as u32).unwrap();
+        Dram::store16(dram, addr, minus_num as u64).unwrap();
         assert_ne!(minus_num, Dram::load_u16(dram, addr).unwrap() as i32);
-        Dram::store16(dram, addr, minus_num as u32).unwrap();
+        Dram::store16(dram, addr, minus_num as u64).unwrap();
         assert_eq!(65494, Dram::load_u16(dram, addr).unwrap());
     }
 
@@ -249,8 +249,8 @@ mod tests {
         };
         let mut addr = 0;
         let mut test_32 = |data: i32| {
-            Dram::store32(dram, addr, data as u32).unwrap();
-            assert_eq!(data as u32, Dram::load32(dram, addr).unwrap());
+            Dram::store32(dram, addr, data as u64).unwrap();
+            assert_eq!(data as u64, Dram::load32(dram, addr).unwrap());
             addr += 2;
         };
